fix: Optional 타입 남용 수정
+ ItemOrder position 필드 타입 Float에서 Double로 수정함.

+++ src/main/java/com/team/routineconnect/dto/ItemResponse.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Getter
 public class ItemResponse {
     private Item item;
-    private Float position;
+    private Double position;
     private Boolean accomplishment;
     private String retrospective;
 }
